+++ src/Integration/Bricks/Main.php
@@ -18,7 +18,6 @@ use WindPress\WindPress\Utils\Common;
 use WindPress\WindPress\Utils\Config;
 
 /**
- * Tested with Bricks version 1.10.3
  * @author Joshua Gugun Siagian <suabahasa@gmail.com>
  */
 class Main implements IntegrationInterface